[TASK-21] Restructure question builders
