+++ src/helpers/questionProperties.js
