+++ src/tests/screens/SurveyDetail/detail.test.js
@@ -7,8 +7,8 @@ import { Router } from 'react-router-dom';
 
 import { DetailsContext } from 'contexts/details';
 import { SurveyStatusProvider } from 'contexts/surveyStatus';
+import buildQuestionList from 'helpers/buildQuestionList';
 import Detail from 'screens/SurveyDetail/detail';
-import buildQuestionList from 'screens/SurveyDetail/QuestionBuilders/buildQuestionList';
 import SurveyDetailResponse from 'tests/fixtures/surveyDetailResponse.json';
 
 describe('When Detail component is mounted', () => {
